Improve kmalloc() and map LAPIC registers into virtual memory
In order to map the LAPIC registers, we need a way to allocate arbitrary
physical pages into the kernel heap. We do this by restructuring
kmalloc() as a wrapper function around the more general alloc(), a
routine which accepts a pointer to a function returning the PMA of the
next page to be mapped into the heap when necessary.