Implement context switching
This giant commit implements a process table and a simple round-robin
scheduler with a simple priority system. Processes are initialized to
run in ring 3 and stack switching is properly configured for privilege
changes.

The system currently registers 3 simple processes for execution. The
main kernel thread is also added to the execution queue and simply halts
repeatedly, serving effectively as a system idler. The idler's process
priority could be dynamically adjusted during runtime for simple
power-management purposes.

Lastly, there are many additions/improvements to code.